Move panel into chronology subdomain, split views by family
Aligns the Panel bounded context with GLIMPSE conventions
(gates/{port}/{adapter}/{subdomain}/{context}/) by moving
gates/web/django/panel.py (3561 lines) into chronology/panel/views/
and splitting it into per-area modules. Panel URL config moves to
chronology/panel/urls.py alongside the existing timetable patterns.
