+++ src/ludamus/gates/web/django/urls.py
@@ -11,9 +11,6 @@ from django.http import JsonResponse
 from django.urls import include, path
 from django.views.decorators.cache import never_cache
 
-from ludamus.gates.web.django import panel
-from ludamus.gates.web.django.chronology.panel.urls import timetable_urlpatterns
-
 if TYPE_CHECKING:
     from django.http import HttpRequest, HttpResponse
     from django.urls import URLPattern, URLResolver
@@ -48,270 +45,13 @@ def healthz(request: HttpRequest) -> JsonResponse:  # noqa: ARG001
     return JsonResponse({"status": "ok"})
 
 
-panel_urlpatterns: list[URLPattern | URLResolver] = [
-    path("", panel.PanelIndexRedirectView.as_view(), name="index"),
-    path("event/<slug:slug>/", panel.EventIndexPageView.as_view(), name="event-index"),
-    path(
-        "event/<slug:slug>/settings/",
-        panel.EventSettingsPageView.as_view(),
-        name="event-settings",
-    ),
-    path(
-        "event/<slug:slug>/settings/proposals/",
-        panel.EventProposalSettingsPageView.as_view(),
-        name="event-proposal-settings",
-    ),
-    path(
-        "event/<slug:slug>/settings/display/",
-        panel.EventDisplaySettingsPageView.as_view(),
-        name="event-display-settings",
-    ),
-    path("event/<slug:slug>/venues/", panel.VenuesPageView.as_view(), name="venues"),
-    path(
-        "event/<slug:slug>/venues/structure/",
-        panel.VenuesStructurePageView.as_view(),
-        name="venues-structure",
-    ),
-    path(
-        "event/<slug:slug>/venues/create/",
-        panel.VenueCreatePageView.as_view(),
-        name="venue-create",
-    ),
-    path(
-        "event/<slug:slug>/venues/<str:venue_slug>/edit/",
-        panel.VenueEditPageView.as_view(),
-        name="venue-edit",
-    ),
-    path(
-        "event/<slug:slug>/venues/<str:venue_slug>/do/delete",
-        panel.VenueDeleteActionView.as_view(),
-        name="venue-delete",
-    ),
-    path(
-        "event/<slug:slug>/venues/<str:venue_slug>/do/duplicate",
-        panel.VenueDuplicatePageView.as_view(),
-        name="venue-duplicate",
-    ),
-    path(
-        "event/<slug:slug>/venues/<str:venue_slug>/do/copy",
-        panel.VenueCopyPageView.as_view(),
-        name="venue-copy",
-    ),
-    path(
-        "event/<slug:slug>/venues/do/reorder",
-        panel.VenueReorderActionView.as_view(),
-        name="venue-reorder",
-    ),
-    path(
-        "event/<slug:slug>/venues/<str:venue_slug>/",
-        panel.VenueDetailPageView.as_view(),
-        name="venue-detail",
-    ),
-    path(
-        "event/<slug:slug>/venues/<str:venue_slug>/areas/create/",
-        panel.AreaCreatePageView.as_view(),
-        name="area-create",
-    ),
-    path(
-        "event/<slug:slug>/venues/<str:venue_slug>/areas/<str:area_slug>/edit/",
-        panel.AreaEditPageView.as_view(),
-        name="area-edit",
-    ),
-    path(
-        "event/<slug:slug>/venues/<str:venue_slug>/areas/<str:area_slug>/do/delete",
-        panel.AreaDeleteActionView.as_view(),
-        name="area-delete",
-    ),
-    path(
-        "event/<slug:slug>/venues/<str:venue_slug>/areas/do/reorder",
-        panel.AreaReorderActionView.as_view(),
-        name="area-reorder",
-    ),
-    path(
-        "event/<slug:slug>/venues/<str:venue_slug>/areas/<str:area_slug>/",
-        panel.AreaDetailPageView.as_view(),
-        name="area-detail",
-    ),
-    path(
-        "event/<slug:slug>/venues/<str:venue_slug>/areas/<str:area_slug>/spaces/create/",
-        panel.SpaceCreatePageView.as_view(),
-        name="space-create",
-    ),
-    path(
-        "event/<slug:slug>/venues/<str:venue_slug>/areas/<str:area_slug>/spaces/"
-        "<str:space_slug>/edit/",
-        panel.SpaceEditPageView.as_view(),
-        name="space-edit",
-    ),
-    path(
-        "event/<slug:slug>/venues/<str:venue_slug>/areas/<str:area_slug>/spaces/"
-        "<str:space_slug>/do/delete",
-        panel.SpaceDeleteActionView.as_view(),
-        name="space-delete",
-    ),
-    path(
-        "event/<slug:slug>/venues/<str:venue_slug>/areas/<str:area_slug>/spaces/do/reorder",
-        panel.SpaceReorderActionView.as_view(),
-        name="space-reorder",
-    ),
-    path(
-        "event/<slug:slug>/cfp/time-slots/",
-        panel.TimeSlotsPageView.as_view(),
-        name="time-slots",
-    ),
-    path(
-        "event/<slug:slug>/cfp/time-slots/create/",
-        panel.TimeSlotCreatePageView.as_view(),
-        name="time-slot-create",
-    ),
-    path(
-        "event/<slug:slug>/cfp/time-slots/<int:pk>/edit/",
-        panel.TimeSlotEditPageView.as_view(),
-        name="time-slot-edit",
-    ),
-    path(
-        "event/<slug:slug>/cfp/time-slots/<int:pk>/do/delete",
-        panel.TimeSlotDeleteActionView.as_view(),
-        name="time-slot-delete",
-    ),
-    path(
-        "event/<slug:slug>/proposals/",
-        panel.ProposalsPageView.as_view(),
-        name="proposals",
-    ),
-    path(
-        "event/<slug:slug>/proposals/create/",
-        panel.ProposalCreatePageView.as_view(),
-        name="proposal-create",
-    ),
-    path(
-        "event/<slug:slug>/proposals/<int:proposal_id>/",
-        panel.ProposalDetailPageView.as_view(),
-        name="proposal-detail",
-    ),
-    path(
-        "event/<slug:slug>/proposals/<int:proposal_id>/edit/",
-        panel.ProposalEditPageView.as_view(),
-        name="proposal-edit",
-    ),
-    path(
-        "event/<slug:slug>/proposals/<int:proposal_id>/do/reject",
-        panel.ProposalRejectActionView.as_view(),
-        name="proposal-reject",
-    ),
-    path(
-        "event/<slug:slug>/proposals/<int:proposal_id>/do/set-facilitators",
-        panel.ProposalSetFacilitatorsActionView.as_view(),
-        name="proposal-set-facilitators",
-    ),
-    path("event/<slug:slug>/cfp/", panel.CFPPageView.as_view(), name="cfp"),
-    path(
-        "event/<slug:slug>/cfp/create/",
-        panel.CFPCreatePageView.as_view(),
-        name="cfp-create",
-    ),
-    path(
-        "event/<slug:slug>/cfp/personal-data/",
-        panel.PersonalDataFieldsPageView.as_view(),
-        name="personal-data-fields",
-    ),
-    path(
-        "event/<slug:slug>/cfp/personal-data/create/",
-        panel.PersonalDataFieldCreatePageView.as_view(),
-        name="personal-data-field-create",
-    ),
-    path(
-        "event/<slug:slug>/cfp/personal-data/<str:field_slug>/edit/",
-        panel.PersonalDataFieldEditPageView.as_view(),
-        name="personal-data-field-edit",
-    ),
-    path(
-        "event/<slug:slug>/cfp/personal-data/<str:field_slug>/do/delete",
-        panel.PersonalDataFieldDeleteActionView.as_view(),
-        name="personal-data-field-delete",
-    ),
-    path(
-        "event/<slug:slug>/cfp/session-fields/",
-        panel.SessionFieldsPageView.as_view(),
-        name="session-fields",
-    ),
-    path(
-        "event/<slug:slug>/cfp/session-fields/create/",
-        panel.SessionFieldCreatePageView.as_view(),
-        name="session-field-create",
-    ),
-    path(
-        "event/<slug:slug>/cfp/session-fields/<str:field_slug>/edit/",
-        panel.SessionFieldEditPageView.as_view(),
-        name="session-field-edit",
-    ),
-    path(
-        "event/<slug:slug>/cfp/session-fields/<str:field_slug>/do/delete",
-        panel.SessionFieldDeleteActionView.as_view(),
-        name="session-field-delete",
-    ),
-    path(
-        "event/<slug:event_slug>/cfp/<str:category_slug>/",
-        panel.CFPEditPageView.as_view(),
-        name="cfp-edit",
-    ),
-    path(
-        "event/<slug:event_slug>/cfp/<str:category_slug>/do/delete",
-        panel.CFPDeleteActionView.as_view(),
-        name="cfp-delete",
-    ),
-    path(
-        "parts/icon-preview/", panel.IconPreviewPartView.as_view(), name="icon-preview"
-    ),
-    path("event/<slug:slug>/tracks/", panel.TracksPageView.as_view(), name="tracks"),
-    path(
-        "event/<slug:slug>/tracks/create/",
-        panel.TrackCreatePageView.as_view(),
-        name="track-create",
-    ),
-    path(
-        "event/<slug:slug>/tracks/<str:track_slug>/edit/",
-        panel.TrackEditPageView.as_view(),
-        name="track-edit",
-    ),
-    path(
-        "event/<slug:slug>/tracks/<str:track_slug>/do/delete",
-        panel.TrackDeleteActionView.as_view(),
-        name="track-delete",
-    ),
-    path(
-        "event/<slug:slug>/facilitators/",
-        panel.FacilitatorsPageView.as_view(),
-        name="facilitators",
-    ),
-    path(
-        "event/<slug:slug>/facilitators/create/",
-        panel.FacilitatorCreatePageView.as_view(),
-        name="facilitator-create",
-    ),
-    path(
-        "event/<slug:slug>/facilitators/merge/",
-        panel.FacilitatorMergePageView.as_view(),
-        name="facilitator-merge",
-    ),
-    path(
-        "event/<slug:slug>/facilitators/<str:facilitator_slug>/",
-        panel.FacilitatorDetailPageView.as_view(),
-        name="facilitator-detail",
-    ),
-    path(
-        "event/<slug:slug>/facilitators/<str:facilitator_slug>/edit/",
-        panel.FacilitatorEditPageView.as_view(),
-        name="facilitator-edit",
-    ),
-    path("event/<slug:slug>/timetable/", include(timetable_urlpatterns)),
-]
-
-
 urlpatterns: list[URLResolver | URLPattern] = [
     path("healthz/", healthz),
     path("", include("ludamus.adapters.web.django.urls", namespace="web")),
-    path("panel/", include((panel_urlpatterns, "panel"), namespace="panel")),
+    path(
+        "panel/",
+        include("ludamus.gates.web.django.chronology.panel.urls", namespace="panel"),
+    ),
     path("admin/", admin.site.urls),
     path("page/", include("django.contrib.flatpages.urls")),
 ]
